backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9
pandas==2.1.4
plotly==5.17.0
google-genai==0.9.0
python-dotenv==1.0.0
numpy==1.24.3
openpyxl==3.1.2
